Admin prod

+++ src/app/pages/admin/classes/classes.component.ts
@@ -48,26 +48,48 @@ export class ClassesComponent implements OnInit, OnDestroy {
 
   }
 
-  removeClass(classId: string) {
-    this.isLoading = true;
-    this.removeClassSub = this.classService.deleteClass(classId).subscribe(
-      response => {
-      if (response.success && response.message && response.data) {
-        this.classes.forEach((cls, index) => {
-          if (cls._id === classId) {
-             this.classes.splice(index, 1);
-          }
-        });
-        this.isLoading = false;
-        this.alertService.success(response.message);
-       }
-    },
-    error => {
-      this.isLoading = false;
-      if (error && error.error && error.error.message) {
-        this.alertService.danger(error.error.message);
+  removeClass(classId: string, title) {
+
+    const isMad = confirm('Are you mad?');
+
+    if (isMad) {
+      const isReallyMad = confirm('Are you really MAD? It will destroy all your boards and results!!!');
+      if (isReallyMad) {
+        const classTitle = prompt('Ohh! you are mad! So enter class name to delete...');
+        if (classTitle === title) {
+
+          this.isLoading = true;
+          this.removeClassSub = this.classService.deleteClass(classId).subscribe(
+            response => {
+              if (response.success && response.message && response.data) {
+                this.classes.forEach((cls, index) => {
+                  if (cls._id === classId) {
+                    this.classes.splice(index, 1);
+                  }
+                });
+                this.isLoading = false;
+                this.alertService.success(response.message);
+              }
+            },
+            error => {
+              this.isLoading = false;
+              if (error && error.error && error.error.message) {
+                this.alertService.danger(error.error.message);
+              }
+            });
+
+        } else {
+
+          this.alertService.warning('Ohh Thank God! Invalid class title');
+
+        }
+
+      } else {
+        this.alertService.success('Thank God, you are not really mad');
       }
-    });
+    } else {
+      this.alertService.success('Thank God, you are not mad');
+    }
   }
 
   editClass(clas: any) {
